Fetch Bamboo version via the API
